% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACtionDB-package.R
\docType{package}
\name{ATACtionDB}
\alias{ATACtionDB}
\title{ATACtionDB: Supplementary data package for the ATACtion package}
\description{
Extenstion of ACTIONet package for single-cell epigenomic dataset
}
\section{Useful links}{


\enumerate{
\item Report bugs at \url{https://github.com/shmohammadi86/ATACtionDB/issues}
}
}

