% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EnhDB_Epimap_hg38}
\alias{EnhDB_Epimap_hg38}
\title{Pre-processed collection of enhancers from the Epimap (hg38 coordinate)}
\format{:
  SimpleGenomicRangesList with one entry for each sample (tissue/celltype)}
\source{
\url{http://compbio.mit.edu/epimap}
}
\usage{
EnhDB_Epimap_hg38
}
\description{
Pre-processed collection of enhancers from the Epimap (hg38 coordinate)
}
\keyword{datasets}
