% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LDblocks}
\alias{LDblocks}
\title{Pre-processed LD blocks}
\format{:
  List of independent LD blocks for AFN, ASN, and EUR in GenomicRanges format}
\source{
\url{https://bitbucket.org/nygcresearch/ldetect-data/src/master/}
}
\usage{
LDblocks
}
\description{
Pre-processed LD blocks
}
\keyword{datasets}
